restaurants all working

+++ waiwai_backend/waiwai_backend_2/db/migrate/20190914031930_create_restaurants.rb
@@ -4,7 +4,7 @@ class CreateRestaurants < ActiveRecord::Migration[5.2]
       t.text :name
       t.text :description
       t.text :image
-      t.float :cost_per_person
+      t.integer :cost_per_person
       t.float :stars
     end
   end
